refactor: extract superclass from interceptors
