+++ src/main/java/nextstep/subway/auth/ui/interceptor/authentication/TokenAuthenticationInterceptor.java
@@ -2,7 +2,6 @@ package nextstep.subway.auth.ui.interceptor.authentication;
 
 import nextstep.subway.auth.application.converter.AuthenticationConverter;
 import nextstep.subway.auth.domain.Authentication;
-import nextstep.subway.auth.domain.AuthenticationToken;
 import nextstep.subway.auth.dto.TokenResponse;
 import nextstep.subway.auth.infrastructure.JwtTokenProvider;
 import nextstep.subway.member.application.CustomUserDetailsService;
@@ -12,28 +11,23 @@ import nextstep.subway.util.ConvertUtils;
 import nextstep.subway.util.HttpResponseUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import java.util.Objects;
 
-public class TokenAuthenticationInterceptor implements HandlerInterceptor {
+public class TokenAuthenticationInterceptor extends AbstractAuthenticationInterceptor {
 
-    private final AuthenticationConverter converter;
-    private final CustomUserDetailsService userDetailsService;
     private final JwtTokenProvider jwtTokenProvider;
 
     public TokenAuthenticationInterceptor(AuthenticationConverter converter, CustomUserDetailsService userDetailsService, JwtTokenProvider jwtTokenProvider) {
-        this.converter = converter;
-        this.userDetailsService = userDetailsService;
+        super(converter, userDetailsService);
         this.jwtTokenProvider = jwtTokenProvider;
     }
 
     @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        final AuthenticationToken authToken = converter.convert(request);
-        final Authentication authentication = authenticate(authToken);
+    protected void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
         Objects.requireNonNull(authentication, "Authentication is null.");
 
         final String payload = ConvertUtils.stringify(toMemberResponse((LoginMember) authentication.getPrincipal()));
@@ -42,25 +36,6 @@ public class TokenAuthenticationInterceptor implements HandlerInterceptor {
         response.setStatus(HttpStatus.CREATED.value());
         response.setContentType(MediaType.APPLICATION_JSON_VALUE);
         HttpResponseUtils.write(response, () -> ConvertUtils.stringify(new TokenResponse(jwtToken)));
-        return false;
-    }
-
-    private Authentication authenticate(AuthenticationToken token) {
-        final String principal = token.getPrincipal();
-        final LoginMember loginMember = userDetailsService.loadUserByUsername(principal);
-        checkAuthentication(loginMember, token);
-
-        return new Authentication(loginMember);
-    }
-
-    private void checkAuthentication(LoginMember loginMember, AuthenticationToken token) {
-        if (loginMember == null) {
-            throw new RuntimeException("No such user in repository.");
-        }
-
-        if (!loginMember.checkPassword(token.getCredentials())) {
-            throw new RuntimeException("Wrong password.");
-        }
     }
 
     private MemberResponse toMemberResponse(LoginMember loginMember) {
